Filters for TableComponent, display all countries on map The filters for the tableComponent are now working properly. However there is an issue where the table get's smaller than the panel it's displayed in if the values are too small in space. Typically this happens when there are only very few movies to be displayed.
Moreover I changed the MovieCollectionConverter for the map so that all countries have a value (0 if there are no movies).
If this change turns out to be a bad idea it should be undone.
Unfortunately now countries with few movies cannot be distinguished from countries with no movies at all.
Choosing less colours in colorAxis seems to help a bit.